pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "leao",email = "lucasleaolima@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"

